Upgrading to sbteclipse 3.0.0 to no longer interfere with eclipse
Until sbeclipse 3.0.0 the genereated eclipse files have used the same
directory for compiled classes as sbt. That lead to a permanent race
when running sbt run on the command line and a lot of classnotfound
exceptions. That is hopefullly resolved with this upgrade as eclipse
stores the classes now in a different directory.